api/requirements.txt
Django==3.2.6
beautifulsoup4==4.9.3
django-restframework==0.0.1
Pillow==8.3.1
djangorestframework==3.12.4
